LayoutTests/http/tests/webcodecs/h264-reordering.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
</head>
<body>
    <script>
let decoder;
promise_test(async () => {
     const response = await fetch("/media-resources/media-source/content/test-fragmented.mp4");
     const buffer = await response.arrayBuffer();
     const avcCOffset = [701, 34];
     const frames = [
         [41, 2193, 25120],
         [125, 27313, 2173],
         [83, 29486, 1348],
         [208, 30834, 1049],
         [166, 31883, 865],
         [291, 32748, 930],
         [250, 33678, 812],
         [375, 34490, 1231],
         [333, 35721, 803],
         [458, 36524, 1564]
    ];

    let frameTimestamps = [];
    decoder =  new VideoDecoder({
        output(frame) {
            frameTimestamps.push(frame.timestamp);
            frame.close();
        },
        error(e) {
            console.log(e);
        }
    });
    decoder.configure({
        codec: 'avc1.4d401e',
        codedWidth: 640,
        codedHeight: 480,
        visibleRect: {x: 0, y: 0, width: 640, height: 480},
        displayWidth: 640,
        displayHeight: 480,
        description: new Uint8Array(buffer, avcCOffset[0], avcCOffset[1])
    });

    chunks = frames.map((frame, i) => new EncodedVideoChunk({type: i == 0 ? 'key' : 'delta', timestamp: frame[0], duration: 1, data: new Uint8Array(buffer, frame[1], frame[2])}));

    for (let i = 0; i < chunks.length; ++i) {
        decoder.decode(chunks[i]);
        if (window.internals && (i === 2 || i === 4 || i === 8)) {
             let counter = 0;
             while (++counter < 100 && internals.hasPendingActivity(decoder))
                 await new Promise(resolve => setTimeout(resolve, 50));
             assert_less_than(counter, 100, "test for " + i);
        }
    }

    await decoder.flush();

    assert_array_equals(frameTimestamps, frames.map(frame => frame[0]).sort((a,b) => a - b), "timestamps are ordered");
}, "H264 reordering");

    const frames = [
	{
		data: new Uint8Array([0, 0, 2, 159, 6, 5, 255, 255, 155, 220, 69, 233, 189, 230, 217, 72, 183, 150, 44, 216, 32, 217, 35, 238, 239, 120, 50, 54, 52, 32, 45, 32, 99, 111, 114, 101, 32, 49, 52, 56, 32, 45, 32, 72, 46, 50, 54, 52, 47, 77, 80, 69, 71, 45, 52, 32, 65, 86, 67, 32, 99, 111, 100, 101, 99, 32, 45, 32, 67, 111, 112, 121, 108, 101, 102, 116, 32, 50, 48, 48, 51, 45, 50, 48, 49, 53, 32, 45, 32, 104, 116, 116, 112, 58, 47, 47, 119, 119, 119, 46, 118, 105, 100, 101, 111, 108, 97, 110, 46, 111, 114, 103, 47, 120, 50, 54, 52, 46, 104, 116, 109, 108, 32, 45, 32, 111, 112, 116, 105, 111, 110, 115, 58, 32, 99, 97, 98, 97, 99, 61, 49, 32, 114, 101, 102, 61, 49, 32, 100, 101, 98, 108, 111, 99, 107, 61, 49, 58, 48, 58, 48, 32, 97, 110, 97, 108, 121, 115, 101, 61, 48, 120, 51, 58, 48, 120, 49, 49, 51, 32, 109, 101, 61, 104, 101, 120, 32, 115, 117, 98, 109, 101, 61, 50, 32, 112, 115, 121, 61, 49, 32, 112, 115, 121, 95, 114, 100, 61, 49, 46, 48, 48, 58, 48, 46, 48, 48, 32, 109, 105, 120, 101, 100, 95, 114, 101, 102, 61, 48, 32, 109, 101, 95, 114, 97, 110, 103, 101, 61, 49, 54, 32, 99, 104, 114, 111, 109, 97, 95, 109, 101, 61, 49, 32, 116, 114, 101, 108, 108, 105, 115, 61, 48, 32, 56, 120, 56, 100, 99, 116, 61, 49, 32, 99, 113, 109, 61, 48, 32, 100, 101, 97, 100, 122, 111, 110, 101, 61, 50, 49, 44, 49, 49, 32, 102, 97, 115, 116, 95, 112, 115, 107, 105, 112, 61, 49, 32, 99, 104, 114, 111, 109, 97, 95, 113, 112, 95, 111, 102, 102, 115, 101, 116, 61, 48, 32, 116, 104, 114, 101, 97, 100, 115, 61, 56, 32, 108, 111, 111, 107, 97, 104, 101, 97, 100, 95, 116, 104, 114, 101, 97, 100, 115, 61, 50, 32, 115, 108, 105, 99, 101, 100, 95, 116, 104, 114, 101, 97, 100, 115, 61, 48, 32, 110, 114, 61, 48, 32, 100, 101, 99, 105, 109, 97, 116, 101, 61, 49, 32, 105, 110, 116, 101, 114, 108, 97, 99, 101, 100, 61, 48, 32, 98, 108, 117, 114, 97, 121, 95, 99, 111, 109, 112, 97, 116, 61, 48, 32, 99, 111, 110, 115, 116, 114, 97, 105, 110, 101, 100, 95, 105, 110, 116, 114, 97, 61, 48, 32, 98, 102, 114, 97, 109, 101, 115, 61, 51, 32, 98, 95, 112, 121, 114, 97, 109, 105, 100, 61, 50, 32, 98, 95, 97, 100, 97, 112, 116, 61, 49, 32, 98, 95, 98, 105, 97, 115, 61, 48, 32, 100, 105, 114, 101, 99, 116, 61, 49, 32, 119, 101, 105, 103, 104, 116, 98, 61, 49, 32, 111, 112, 101, 110, 95, 103, 111, 112, 61, 48, 32, 119, 101, 105, 103, 104, 116, 112, 61, 49, 32, 107, 101, 121, 105, 110, 116, 61, 50, 53, 48, 32, 107, 101, 121, 105, 110, 116, 95, 109, 105, 110, 61, 50, 53, 32, 115, 99, 101, 110, 101, 99, 117, 116, 61, 52, 48, 32, 105, 110, 116, 114, 97, 95, 114, 101, 102, 114, 101, 115, 104, 61, 48, 32, 114, 99, 95, 108, 111, 111, 107, 97, 104, 101, 97, 100, 61, 49, 48, 32, 114, 99, 61, 99, 114, 102, 32, 109, 98, 116, 114, 101, 101, 61, 49, 32, 99, 114, 102, 61, 50, 52, 46, 48, 32, 113, 99, 111, 109, 112, 61, 48, 46, 54, 48, 32, 113, 112, 109, 105, 110, 61, 48, 32, 113, 112, 109, 97, 120, 61, 54, 57, 32, 113, 112, 115, 116, 101, 112, 61, 52, 32, 105, 112, 95, 114, 97, 116, 105, 111, 61, 49, 46, 52, 48, 32, 97, 113, 61, 49, 58, 49, 46, 48, 48, 0, 128, 0, 0, 15, 44, 101, 136, 132, 0, 55, 255, 210, 120, 30, 239, 237, 83, 148, 112, 49, 63, 239, 115, 35, 149, 179, 16, 147, 28, 208, 44, 55, 32, 129, 99, 4, 72, 16, 0, 0, 3, 0, 0, 3, 0, 26, 100, 113, 116, 137, 6, 98, 103, 214, 84, 128, 133, 189, 113, 178, 8, 200, 36, 177, 246, 245, 60, 19, 122, 225, 208, 67, 54, 133, 18, 234, 93, 198, 97, 164, 249, 185, 200, 246, 108, 43, 159, 64, 119, 205, 159, 246, 188, 149, 65, 236, 57, 217, 139, 248, 33, 135, 91, 225, 61, 220, 24, 15, 209, 173, 221, 125, 183, 83, 58, 255, 250, 138, 203, 20, 113, 101, 217, 136, 54, 172, 144, 71, 116, 165, 151, 32, 205, 231, 91, 170, 2, 7, 213, 193, 188, 159, 240, 159, 181, 255, 221, 119, 97, 101, 75, 40, 28, 120, 206, 64, 3, 231, 250, 239, 23, 158, 98, 217, 46, 117, 155, 8, 162, 224, 94, 149, 255, 54, 232, 83, 67, 194, 6, 126, 238, 227, 204, 169, 255, 255, 202, 255, 191, 29, 138, 98, 119, 232, 210, 153, 0, 12, 176, 21, 99, 17, 66, 182, 67, 118, 72, 242, 177, 27, 132, 108, 216, 160, 72, 10, 184, 200, 34, 147, 127, 244, 231, 81, 70, 82, 180, 29, 213, 12, 83, 252, 178, 136, 106, 190, 249, 43, 105, 252, 141, 13, 190, 57, 43, 5, 235, 93, 42, 106, 10, 138, 127, 135, 92, 189, 214, 226, 126, 158, 34, 38, 19, 35, 84, 55, 107, 131, 219, 92, 234, 16, 26, 117, 141, 217, 16, 217, 52, 7, 100, 251, 31, 66, 156, 176, 96, 108, 36, 252, 152, 234, 123, 215, 82, 90, 96, 212, 194, 181, 169, 187, 47, 13, 125, 232, 63, 67, 118, 22, 133, 123, 195, 123, 152, 174, 202, 169, 12, 72, 176, 83, 254, 215, 74, 79, 150, 210, 114, 27, 88, 12, 200, 71, 241, 103, 202, 155, 203, 171, 185, 51, 19, 19, 224, 63, 176, 92, 179, 229, 89, 178, 61, 110, 174, 120, 91, 68, 188, 238, 150, 160, 10, 147, 20, 176, 80, 116, 215, 230, 1, 151, 38, 69, 234, 148, 155, 115, 164, 164, 199, 83, 217, 168, 140, 87, 14, 126, 36, 222, 196, 46, 226, 65, 171, 46, 78, 26, 221, 89, 172, 196, 48, 111, 211, 140, 73, 202, 161, 58, 103, 46, 234, 18, 250, 155, 117, 8, 250, 196, 166, 1, 107, 65, 30, 48, 34, 32, 234, 211, 234, 28, 30, 124, 31, 142, 117, 205, 179, 187, 146, 132, 219, 173, 11, 11, 1, 175, 44, 144, 22, 172, 6, 42, 68, 125, 149, 250, 135, 194, 64, 43, 175, 198, 150, 19, 110, 48, 96, 17, 30, 114, 95, 244, 3, 213, 136, 223, 49, 85, 171, 48, 32, 201, 190, 90, 140, 117, 66, 8, 144, 41, 112, 39, 46, 94, 239, 150, 216, 33, 28, 232, 222, 37, 73, 251, 42, 102, 43, 97, 158, 245, 181, 129, 220, 72, 185, 77, 231, 48, 188, 236, 45, 118, 223, 73, 36, 90, 145, 166, 157, 217, 157, 152, 120, 45, 75, 120, 49, 211, 122, 3, 172, 0, 91, 173, 89, 216, 213, 194, 226, 165, 70, 208, 244, 26, 227, 15, 92, 228, 80, 122, 213, 92, 123, 241, 40, 158, 232, 252, 45, 103, 178, 27, 50, 207, 90, 90, 76, 40, 79, 75, 74, 63, 171, 229, 220, 0, 61, 249, 55, 2, 32, 141, 79, 5, 26, 109, 93, 143, 254, 66, 8, 245, 222, 184, 32, 123, 242, 223, 165, 229, 130, 125, 162, 111, 245, 97, 151, 219, 224, 3, 66, 173, 189, 234, 41, 217, 234, 103, 145, 129, 30, 216, 32, 9, 95, 2, 187, 233, 110, 205, 27, 125, 173, 163, 76, 122, 120, 24, 150, 69, 171, 45, 32, 69, 192, 129, 85, 6, 100, 236, 59, 227, 55, 90, 74, 210, 49, 191, 37, 59, 251, 91, 128, 46, 221, 129, 253, 33, 221, 194, 31, 120, 44, 10, 211, 225, 218, 1, 75, 46, 39, 62, 48, 119, 50, 104, 19, 217, 34, 71, 140, 244, 220, 66, 27, 202, 224, 177, 114, 195, 71, 219, 248, 155, 235, 33, 188, 0, 233, 201, 184, 71, 123, 209, 156, 23, 174, 63, 176, 43, 196, 194, 204, 162, 231, 167, 157, 103, 244, 24, 38, 229, 174, 182, 111, 168, 120, 87, 186, 137, 232, 4, 68, 78, 138, 208, 137, 215, 134, 246, 228, 191, 254, 210, 55, 144, 139, 199, 166, 186, 198, 161, 141, 212, 149, 174, 236, 201, 247, 198, 70, 146, 20, 216, 40, 22, 4, 186, 120, 135, 32, 190, 123, 92, 104, 108, 63, 154, 139, 201, 181, 183, 239, 193, 203, 194, 192, 66, 207, 177, 116, 243, 177, 118, 144, 161, 14, 159, 24, 110, 52, 41, 104, 78, 250, 154, 234, 61, 8, 2, 143, 169, 198, 155, 250, 9, 102, 180, 155, 242, 8, 118, 31, 233, 6, 84, 157, 34, 210, 150, 198, 43, 139, 19, 68, 45, 195, 17, 51, 163, 248, 51, 38, 169, 156, 44, 174, 188, 187, 121, 46, 61, 156, 112, 31, 56, 42, 236, 27, 160, 115, 121, 65, 247, 184, 85, 11, 199, 233, 122, 223, 104, 92, 159, 129, 45, 15, 1, 105, 34, 236, 190, 54, 76, 144, 193, 248, 41, 171, 232, 221, 19, 30, 250, 255, 113, 162, 224, 28, 244, 154, 253, 98, 102, 185, 204, 225, 223, 94, 205, 107, 187, 86, 99, 58, 198, 155, 87, 89, 222, 191, 142, 47, 67, 112, 248, 20, 146, 68, 153, 199, 182, 131, 9, 58, 14, 133, 144, 183, 170, 55, 25, 233, 51, 58, 218, 100, 236, 41, 204, 92, 253, 54, 186, 58, 216, 203, 214, 75, 19, 149, 222, 136, 180, 43, 144, 232, 68, 105, 183, 65, 252, 194, 212, 51, 133, 97, 99, 138, 214, 120, 113, 112, 36, 102, 126, 126, 148, 62, 254, 83, 125, 189, 242, 94, 189, 231, 94, 235, 127, 106, 19, 69, 129, 34, 129, 176, 69, 201, 50, 179, 180, 226, 191, 46, 37, 8, 82, 16, 114, 137, 141, 202, 189, 161, 227, 186, 130, 20, 109, 87, 130, 32, 211, 132, 85, 208, 100, 53, 67, 67, 111, 141, 136, 138, 10, 158, 240, 71, 60, 112, 53, 123, 87, 17, 46, 244, 58, 165, 253, 243, 48, 86, 231, 82, 68, 177, 158, 142, 13, 255, 209, 225, 193, 182, 220, 197, 121, 3, 138, 206, 37, 242, 203, 135, 237, 180, 105, 135, 76, 157, 17, 175, 186, 107, 142, 115, 214, 110, 100, 28, 220, 139, 32, 183, 126, 154, 140, 36, 152, 102, 153, 218, 174, 178, 116, 155, 245, 254, 8, 118, 36, 146, 101, 84, 196, 88, 71, 226, 45, 124, 158, 70, 253, 172, 2, 9, 240, 194, 173, 32, 20, 201, 242, 142, 211, 174, 142, 219, 246, 121, 19, 162, 0, 83, 30, 217, 211, 143, 250, 64, 136, 146, 192, 161, 88, 191, 251, 234, 145, 62, 187, 178, 15, 48, 2, 4, 39, 89, 88, 99, 86, 193, 18, 76, 223, 138, 197, 253, 165, 101, 146, 174, 24, 245, 71, 232, 254, 209, 214, 91, 180, 193, 169, 44, 83, 242, 27, 27, 128, 5, 228, 192, 201, 57, 133, 223, 227, 96, 101, 239, 167, 14, 148, 72, 18, 116, 170, 141, 85, 223, 88, 107, 102, 172, 236, 179, 206, 30, 92, 3, 184, 179, 189, 7, 135, 87, 111, 229, 167, 62, 140, 156, 188, 73, 187, 194, 244, 175, 96, 36, 150, 154, 43, 12, 142, 104, 190, 151, 207, 252, 88, 252, 80, 15, 150, 247, 246, 53, 134, 70, 226, 216, 84, 29, 100, 29, 165, 217, 143, 58, 245, 166, 71, 126, 75, 189, 138, 196, 37, 84, 79, 3, 249, 64, 16, 52, 71, 183, 36, 48, 119, 253, 253, 64, 43, 13, 224, 80, 18, 179, 114, 98, 100, 76, 83, 10, 202, 129, 101, 144, 84, 23, 189, 189, 214, 136, 9, 175, 219, 154, 70, 205, 117, 23, 98, 229, 109, 13, 251, 215, 195, 201, 77, 107, 48, 42, 6, 102, 143, 218, 135, 109, 97, 45, 26, 158, 19, 125, 106, 113, 146, 254, 89, 141, 193, 94, 162, 106, 81, 114, 61, 254, 83, 170, 47, 161, 110, 158, 144, 85, 154, 166, 2, 149, 29, 139, 245, 139, 71, 32, 178, 83, 82, 104, 128, 188, 129, 133, 90, 35, 93, 185, 63, 77, 164, 117, 22, 64, 63, 179, 114, 211, 51, 104, 120, 141, 10, 74, 67, 29, 19, 173, 26, 5, 46, 94, 211, 30, 119, 174, 240, 42, 123, 212, 72, 56, 36, 23, 69, 20, 226, 41, 197, 118, 174, 171, 72, 201, 225, 248, 28, 164, 94, 124, 50, 64, 109, 177, 238, 123, 50, 58, 253, 69, 83, 58, 192, 16, 135, 151, 124, 195, 123, 132, 52, 71, 154, 27, 12, 148, 213, 140, 202, 47, 162, 29, 248, 181, 75, 198, 255, 174, 64, 144, 175, 163, 107, 188, 143, 136, 88, 216, 128, 226, 178, 245, 61, 86, 96, 197, 40, 172, 141, 138, 180, 191, 239, 227, 180, 22, 84, 187, 127, 108, 31, 196, 65, 185, 148, 164, 247, 29, 151, 150, 41, 175, 108, 185, 221, 145, 163, 71, 53, 107, 103, 4, 170, 60, 81, 112, 132, 46, 202, 158, 33, 202, 113, 91, 7, 116, 116, 143, 0, 233, 46, 237, 83, 168, 34, 125, 117, 25, 140, 255, 200, 115, 37, 189, 110, 91, 94, 203, 159, 97, 152, 233, 70, 208, 4, 1, 50, 153, 176, 249, 76, 176, 41, 18, 150, 168, 166, 113, 159, 143, 157, 237, 194, 73, 157, 32, 244, 191, 74, 22, 167, 43, 121, 237, 83, 85, 196, 227, 80, 172, 104, 95, 228, 245, 233, 31, 26, 177, 203, 64, 143, 61, 252, 79, 177, 163, 76, 180, 95, 158, 92, 96, 220, 7, 19, 75, 254, 147, 211, 163, 103, 177, 51, 68, 8, 66, 85, 142, 2, 95, 45, 216, 146, 43, 205, 7, 246, 154, 254, 119, 101, 154, 227, 6, 165, 22, 4, 82, 44, 42, 70, 19, 171, 4, 124, 157, 80, 25, 210, 219, 156, 136, 147, 182, 255, 91, 106, 73, 206, 90, 120, 187, 50, 193, 126, 32, 238, 4, 32, 87, 152, 26, 143, 22, 75, 0, 250, 98, 41, 212, 148, 230, 37, 194, 170, 147, 238, 110, 78, 204, 229, 182, 14, 163, 205, 210, 236, 37, 88, 137, 121, 120, 49, 217, 184, 4, 41, 121, 170, 51, 88, 88, 63, 198, 152, 220, 106, 113, 129, 162, 116, 115, 116, 7, 111, 194, 157, 54, 20, 181, 221, 146, 174, 12, 254, 31, 115, 247, 188, 92, 80, 252, 34, 58, 104, 121, 21, 224, 56, 50, 55, 204, 222, 137, 188, 188, 216, 31, 8, 46, 66, 209, 142, 235, 85, 187, 214, 196, 108, 154, 149, 62, 198, 159, 217, 157, 13, 43, 151, 227, 37, 232, 171, 197, 81, 165, 56, 236, 88, 120, 90, 16, 160, 125, 167, 216, 203, 125, 30, 166, 65, 9, 72, 115, 82, 252, 198, 194, 7, 6, 192, 17, 99, 45, 201, 110, 107, 73, 239, 195, 156, 222, 222, 221, 74, 237, 224, 12, 207, 197, 247, 236, 9, 221, 50, 71, 9, 234, 86, 181, 47, 91, 91, 75, 79, 134, 168, 6, 67, 168, 95, 33, 74, 131, 243, 144, 118, 205, 6, 15, 141, 103, 46, 111, 32, 54, 157, 249, 85, 230, 183, 8, 88, 151, 231, 77, 154, 144, 112, 190, 136, 246, 89, 180, 186, 50, 163, 172, 128, 112, 237, 8, 244, 151, 133, 71, 149, 222, 42, 2, 225, 83, 133, 202, 176, 141, 163, 93, 98, 39, 233, 41, 190, 190, 181, 61, 160, 56, 115, 153, 209, 192, 138, 107, 37, 229, 80, 241, 151, 20, 73, 133, 250, 67, 73, 138, 245, 80, 87, 149, 91, 47, 158, 248, 53, 69, 51, 58, 135, 146, 127, 64, 199, 235, 33, 196, 100, 247, 203, 133, 184, 110, 180, 14, 242, 35, 151, 21, 209, 233, 110, 121, 24, 153, 147, 29, 64, 74, 190, 218, 137, 209, 230, 185, 251, 218, 233, 202, 209, 4, 226, 232, 165, 207, 37, 60, 201, 224, 7, 229, 50, 168, 134, 158, 37, 215, 110, 238, 214, 119, 163, 105, 240, 223, 133, 233, 147, 228, 165, 211, 19, 69, 194, 251, 102, 13, 130, 231, 97, 207, 26, 141, 192, 140, 125, 191, 205, 29, 80, 126, 225, 125, 244, 20, 144, 214, 128, 1, 130, 10, 51, 178, 244, 201, 247, 87, 196, 137, 85, 208, 66, 245, 248, 76, 182, 217, 253, 171, 238, 250, 183, 104, 252, 238, 45, 214, 196, 83, 164, 158, 205, 140, 67, 91, 223, 18, 74, 68, 25, 78, 192, 112, 223, 56, 228, 199, 115, 63, 254, 175, 127, 95, 186, 64, 3, 76, 247, 241, 240, 195, 169, 126, 255, 174, 127, 225, 200, 97, 27, 158, 150, 176, 214, 3, 165, 121, 77, 143, 68, 4, 77, 110, 199, 100, 235, 212, 171, 209, 179, 247, 59, 111, 196, 143, 138, 173, 195, 138, 53, 145, 156, 39, 34, 232, 21, 14, 52, 75, 171, 242, 95, 47, 242, 216, 216, 152, 133, 32, 116, 185, 46, 62, 77, 164, 1, 0, 156, 0, 7, 243, 153, 13, 23, 47, 10, 211, 16, 168, 11, 210, 9, 130, 206, 108, 79, 154, 82, 53, 212, 212, 1, 26, 185, 124, 185, 135, 142, 82, 90, 161, 55, 204, 182, 44, 198, 159, 171, 46, 4, 115, 142, 25, 43, 218, 56, 9, 138, 11, 7, 161, 227, 118, 34, 189, 127, 43, 216, 7, 221, 71, 254, 105, 217, 63, 118, 100, 19, 246, 180, 185, 208, 133, 71, 179, 71, 147, 155, 57, 228, 194, 143, 149, 210, 130, 111, 235, 195, 152, 187, 185, 156, 153, 170, 124, 51, 117, 87, 56, 149, 154, 24, 40, 12, 185, 203, 19, 214, 138, 99, 233, 28, 90, 209, 218, 154, 249, 137, 249, 70, 140, 64, 101, 171, 71, 22, 108, 75, 222, 67, 70, 230, 133, 171, 136, 30, 224, 88, 31, 20, 60, 212, 63, 20, 247, 147, 64, 3, 101, 110, 107, 46, 151, 36, 108, 229, 211, 244, 194, 248, 253, 129, 66, 75, 89, 54, 80, 62, 255, 81, 197, 183, 67, 21, 62, 41, 66, 168, 38, 94, 133, 33, 70, 217, 172, 77, 206, 49, 12, 109, 82, 177, 239, 24, 123, 250, 202, 41, 166, 223, 10, 70, 103, 239, 66, 128, 3, 178, 120, 240, 232, 91, 94, 130, 38, 133, 157, 186, 168, 255, 214, 204, 108, 112, 148, 100, 210, 176, 53, 27, 236, 199, 31, 127, 118, 242, 215, 124, 190, 25, 76, 145, 99, 97, 92, 227, 217, 152, 127, 139, 250, 169, 248, 128, 45, 60, 99, 250, 231, 220, 138, 107, 15, 135, 217, 203, 223, 217, 84, 203, 222, 17, 83, 245, 197, 79, 26, 153, 228, 155, 30, 149, 186, 61, 215, 11, 34, 120, 51, 147, 110, 184, 153, 106, 87, 213, 117, 217, 244, 74, 146, 102, 51, 244, 33, 138, 206, 74, 120, 159, 134, 254, 227, 59, 221, 179, 141, 182, 90, 4, 115, 234, 220, 128, 217, 108, 57, 190, 162, 116, 87, 57, 104, 163, 153, 160, 244, 75, 20, 222, 182, 152, 0, 125, 253, 92, 173, 239, 132, 109, 208, 14, 116, 194, 236, 179, 22, 28, 117, 25, 213, 21, 215, 141, 98, 186, 137, 166, 97, 222, 177, 255, 236, 176, 110, 49, 44, 253, 132, 153, 12, 46, 245, 23, 77, 111, 28, 133, 145, 75, 149, 226, 94, 253, 230, 31, 71, 89, 71, 146, 173, 64, 109, 142, 23, 127, 204, 32, 115, 77, 87, 87, 204, 47, 113, 34, 138, 11, 16, 54, 135, 116, 79, 137, 6, 221, 129, 252, 158, 66, 126, 204, 212, 23, 43, 69, 175, 215, 133, 157, 187, 78, 226, 246, 57, 235, 112, 3, 96, 118, 141, 98, 46, 155, 51, 225, 134, 203, 73, 12, 224, 177, 121, 85, 225, 160, 201, 244, 184, 247, 115, 178, 166, 12, 131, 24, 235, 68, 243, 75, 92, 153, 162, 49, 167, 177, 26, 160, 5, 250, 131, 73, 68, 172, 98, 214, 67, 100, 226, 231, 168, 161, 81, 247, 233, 156, 139, 192, 173, 125, 191, 168, 123, 159, 0, 174, 173, 68, 75, 165, 171, 224, 154, 38, 71, 204, 15, 91, 154, 188, 71, 44, 17, 142, 164, 227, 167, 247, 51, 4, 219, 160, 190, 246, 176, 91, 216, 212, 70, 114, 211, 15, 96, 83, 110, 245, 16, 219, 137, 217, 87, 63, 13, 206, 52, 222, 167, 225, 118, 119, 50, 131, 86, 120, 24, 2, 156, 220, 100, 240, 159, 245, 25, 201, 218, 42, 81, 203, 43, 87, 248, 181, 90, 70, 98, 98, 51, 124, 97, 214, 98, 161, 43, 123, 162, 30, 191, 134, 12, 104, 47, 150, 45, 64, 85, 65, 203, 213, 28, 106, 202, 109, 115, 72, 197, 131, 213, 244, 166, 203, 249, 153, 80, 42, 196, 206, 212, 218, 46, 32, 132, 250, 119, 99, 15, 118, 247, 255, 114, 22, 139, 228, 251, 185, 185, 80, 40, 181, 29, 164, 68, 227, 135, 160, 66, 223, 90, 117, 115, 79, 28, 202, 32, 225, 61, 196, 203, 45, 8, 141, 62, 106, 99, 169, 221, 36, 13, 124, 18, 56, 230, 180, 195, 114, 245, 212, 208, 103, 113, 100, 253, 72, 130, 171, 95, 21, 202, 125, 49, 12, 116, 131, 48, 31, 163, 138, 19, 234, 32, 156, 196, 31, 165, 136, 95, 170, 180, 71, 191, 157, 76, 41, 173, 17, 117, 13, 119, 189, 98, 206, 103, 18, 209, 5, 100, 118, 46, 205, 114, 140, 117, 56, 128, 20, 136, 137, 18, 254, 144, 243, 128, 71, 55, 100, 251, 181, 57, 217, 110, 135, 84, 47, 117, 216, 96, 149, 31, 94, 147, 66, 218, 10, 183, 168, 141, 75, 177, 13, 141, 47, 201, 19, 77, 218, 175, 114, 249, 227, 23, 189, 7, 119, 20, 95, 28, 209, 34, 188, 204, 51, 163, 13, 42, 248, 52, 31, 230, 11, 237, 38, 121, 204, 185, 132, 10, 52, 7, 62, 184, 250, 68, 195, 151, 66, 76, 124, 110, 230, 135, 101, 185, 98, 112, 233, 245, 114, 105, 14, 66, 77, 191, 241, 241, 23, 135, 28, 235, 184, 33, 84, 222, 166, 34, 160, 160, 95, 41, 166, 21, 207, 61, 9, 64, 241, 159, 229, 87, 178, 5, 132, 218, 146, 154, 246, 50, 223, 66, 35, 57, 169, 68, 132, 153, 116, 30, 29, 119, 208, 253, 178, 241, 247, 130, 199, 45, 195, 211, 135, 75, 32, 183, 88, 19, 68, 225, 251, 196, 142, 252, 122, 162, 172, 111, 93, 226, 138, 66, 93, 48, 202, 91, 143, 248, 203, 103, 151, 66, 172, 114, 61, 0, 22, 163, 54, 229, 35, 234, 106, 13, 44, 161, 23, 152, 164, 36, 168, 240, 19, 219, 84, 248, 244, 116, 146, 136, 254, 221, 235, 119, 212, 250, 187, 39, 156, 195, 182, 41, 179, 37, 132, 26, 89, 63, 169, 121, 35, 94, 100, 112, 118, 231, 207, 45, 103, 170, 47, 93, 42, 17, 115, 244, 185, 96, 238, 80, 236, 9, 208, 28, 143, 132, 34, 148, 158, 56, 88, 129, 186, 49, 98, 195, 175, 72, 189, 224, 222, 237, 205, 49, 96, 114, 45, 57, 182, 121, 51, 95, 102, 249, 23, 26, 77, 160, 226, 253, 114, 191, 195, 163, 244, 135, 96, 70, 218, 245, 198, 248, 213, 135, 233, 180, 156, 82, 155, 148, 103, 136, 125, 98, 148, 193, 28, 53, 12, 204, 250, 218, 88, 213, 50, 42, 195, 182, 71, 174, 115, 37, 250, 158, 174, 25, 61, 130, 114, 45, 41, 136, 64, 89, 121, 100, 249, 155, 109, 72, 196, 8, 237, 53, 232, 183, 245, 139, 227, 123, 77, 97, 216, 141, 254, 43, 175, 93, 183, 233, 208, 110, 191, 8, 148, 193, 52, 14, 173, 17, 245, 223, 25, 243, 5, 244, 145, 2, 75, 147, 144, 188, 128, 0, 1, 26, 48, 11, 195, 209, 65, 82, 42, 246, 35, 16, 242, 151, 60, 144, 235, 172, 1, 136, 241, 132, 236, 108, 81, 48, 89, 104, 45, 18, 187, 151, 68, 33, 18, 137, 45, 130, 181, 40, 1, 154, 250, 239, 196, 86, 239, 97, 210, 207, 83, 228, 40, 18, 131, 157, 83, 221, 198, 27, 41, 132, 142, 183, 63, 193, 235, 240, 172, 55, 52, 93, 103, 165, 4, 85, 195, 113, 201, 22, 229, 206, 205, 238, 15, 254, 122, 219, 103, 26, 73, 19, 111, 186, 145, 207, 19, 243, 41, 235, 7, 131, 197, 136, 254, 91, 39, 247, 114, 81, 239, 228, 41, 225, 23, 240, 195, 22, 205, 24, 217, 89, 172, 7, 91, 232, 36, 28, 255, 146, 115, 218, 121, 27, 250, 65, 98, 2, 230, 215, 21, 184, 205, 124, 198, 205, 241, 54, 55, 30, 145, 176, 41, 27, 54, 136, 34, 78, 28, 233, 62, 138, 102, 240, 0, 18, 153, 106, 68, 115, 174, 183, 111, 223, 235, 149, 72, 30, 229, 203, 61, 63, 3, 81, 175, 36, 203, 107, 104, 91, 83, 164, 16, 74, 136, 244, 90, 64, 135, 234, 11, 35, 15, 127, 82, 150, 85, 65, 25, 83, 126, 54, 12, 52, 92, 208, 159, 94, 230, 37, 32, 192, 125, 242, 155, 64, 83, 229, 164, 180, 227, 5, 50, 223, 52, 123, 158, 3, 32, 114, 135, 37, 54, 185, 144, 244, 37, 165, 111, 6, 95, 135, 206, 132, 174, 126, 116, 21, 80, 81, 93, 217, 53, 213, 226, 112, 255, 45, 207, 139, 140, 150, 37, 251, 90, 181, 98, 118, 33, 173, 77, 114, 193, 24, 89, 9, 102, 66, 106, 69, 182, 93, 59, 124, 71, 233, 115, 110, 227, 92, 83, 219, 47, 194, 235, 248, 190, 226, 180, 2, 243, 57, 219, 112, 191, 221, 72, 145, 40, 211, 206, 134, 64, 0, 25, 45, 197, 28, 80, 157, 22, 252, 45, 144, 90, 75, 123, 26, 115, 146, 195, 40, 119, 214, 235, 75, 82, 246, 70, 247, 35, 81, 112, 233, 94, 192, 97, 240, 213, 76, 28, 63, 27, 71, 60, 242, 219, 179, 119, 213, 125, 115, 124, 241, 59, 143, 219, 128, 72, 248, 67, 48, 160, 74, 149, 147, 98, 52, 23, 43, 128, 53, 87, 105, 94, 234, 6, 21, 122, 120, 157, 245, 158, 61, 242, 129, 125, 77, 231, 147, 174, 185, 202, 30, 138, 76, 228, 98, 106, 234, 136, 83, 146, 147, 107, 53, 117, 97, 20, 35, 156, 247, 49, 114, 21, 186, 162, 52, 91, 37, 244, 213, 203, 213, 52, 59, 169, 54, 31, 104, 89, 8, 48, 213, 64, 38, 159, 155, 219, 111, 174, 62, 80, 144, 253, 77, 255, 158, 108, 39, 162, 15, 127, 28, 205, 96, 1, 174, 40, 209, 79, 192, 227, 129]),
		timestamp: 0,
	},
	{
		data: new Uint8Array([0, 0, 1, 2, 65, 154, 36, 24, 134, 127, 254, 140, 176, 24, 231, 28, 204, 254, 48, 5, 106, 159, 118, 61, 175, 66, 199, 51, 22, 234, 186, 125, 139, 77, 192, 92, 177, 62, 30, 99, 233, 196, 14, 189, 175, 218, 151, 97, 255, 5, 54, 151, 207, 27, 184, 17, 86, 144, 243, 72, 170, 68, 2, 84, 143, 97, 203, 179, 239, 71, 249, 28, 79, 217, 41, 9, 195, 120, 51, 164, 67, 198, 198, 14, 56, 65, 212, 230, 252, 215, 255, 176, 62, 11, 129, 97, 220, 73, 143, 193, 216, 227, 77, 102, 224, 185, 24, 69, 174, 123, 243, 91, 86, 227, 52, 239, 243, 254, 42, 34, 119, 168, 214, 70, 249, 247, 178, 18, 134, 62, 163, 174, 142, 238, 43, 53, 149, 72, 82, 252, 115, 147, 180, 170, 157, 63, 54, 128, 145, 181, 208, 106, 170, 219, 236, 187, 163, 0, 75, 78, 208, 228, 36, 89, 209, 161, 78, 181, 98, 107, 212, 22, 159, 68, 51, 199, 181, 74, 17, 241, 210, 96, 99, 133, 81, 200, 82, 221, 119, 37, 62, 200, 113, 227, 160, 232, 81, 34, 185, 57, 48, 81, 252, 34, 48, 127, 87, 231, 64, 142, 91, 88, 176, 26, 120, 167, 38, 32, 192, 116, 180, 1, 34, 5, 22, 186, 73, 252, 235, 155, 254, 95, 23, 6, 131, 135, 143, 50, 90, 66, 206, 3, 218, 214, 215, 251, 42, 55, 178, 109, 152, 53, 136, 102, 147, 200, 41, 195, 191, 107, 200, 128]),
		timestamp: 133333,
	},
	{
		data: new Uint8Array([0, 0, 0, 41, 65, 158, 66, 66, 21, 255, 0, 60, 208, 98, 140, 94, 21, 78, 193, 147, 226, 1, 103, 247, 136, 247, 161, 201, 111, 141, 254, 193, 127, 25, 214, 194, 123, 116, 85, 181, 137, 155, 108, 3, 143]),
		timestamp: 66666,
	},
	{
		data: new Uint8Array([0, 0, 0, 11, 1, 158, 97, 68, 39, 255, 0, 0, 3, 1, 73]),
		timestamp: 33333,
	},
	{
		data: new Uint8Array([0, 0, 0, 13, 1, 158, 99, 68, 39, 255, 0, 15, 141, 25, 60, 137, 217]),
		timestamp: 100000,
	},
	{
		data: new Uint8Array([0, 0, 1, 96, 65, 154, 104, 52, 76, 67, 63, 254, 140, 176, 4, 193, 68, 141, 0, 86, 165, 90, 147, 91, 185, 182, 32, 120, 68, 15, 152, 62, 128, 67, 38, 185, 62, 175, 62, 146, 186, 11, 61, 234, 206, 134, 58, 3, 7, 109, 182, 118, 219, 189, 159, 152, 85, 129, 147, 186, 186, 96, 34, 44, 1, 230, 1, 209, 145, 95, 36, 127, 96, 8, 86, 133, 176, 7, 225, 166, 90, 15, 123, 238, 36, 52, 94, 30, 230, 58, 90, 188, 52, 203, 59, 120, 165, 103, 195, 138, 198, 200, 177, 14, 192, 95, 111, 241, 29, 58, 132, 21, 148, 241, 52, 212, 94, 26, 164, 68, 163, 210, 175, 73, 143, 245, 238, 50, 245, 211, 255, 102, 142, 49, 150, 116, 16, 152, 178, 169, 168, 99, 197, 11, 61, 180, 244, 55, 108, 79, 53, 66, 76, 189, 86, 68, 214, 54, 76, 241, 25, 204, 46, 11, 199, 161, 136, 241, 90, 173, 145, 141, 28, 14, 118, 19, 67, 166, 154, 19, 227, 209, 95, 227, 61, 101, 141, 60, 52, 157, 81, 101, 153, 12, 167, 232, 251, 195, 223, 7, 187, 216, 71, 169, 146, 218, 122, 66, 165, 92, 32, 224, 51, 82, 13, 111, 202, 19, 174, 44, 41, 109, 26, 34, 241, 29, 177, 217, 218, 109, 70, 107, 21, 222, 183, 54, 254, 149, 207, 234, 87, 123, 7, 4, 223, 39, 66, 1, 252, 227, 83, 158, 185, 42, 143, 56, 246, 133, 35, 219, 54, 178, 78, 137, 98, 84, 93, 30, 84, 201, 184, 146, 123, 246, 81, 226, 30, 144, 177, 241, 242, 179, 97, 38, 111, 198, 220, 62, 90, 210, 25, 65, 14, 81, 215, 28, 15, 130, 254, 87, 182, 252, 27, 131, 191, 99, 4, 111, 40, 193, 245, 179, 208, 162, 8, 225, 69, 42, 98, 248, 167, 219, 204, 229, 215, 55, 10, 13, 169, 149, 86, 39, 74, 206, 120, 215, 228, 159, 94, 239, 1, 206, 180, 145, 233, 228, 192, 8, 247, 229, 232, 28, 41, 78, 111, 65]),
		timestamp: 266666,
	},
	{
		data: new Uint8Array([0, 0, 0, 45, 65, 158, 134, 69, 17, 44, 43, 255, 0, 67, 121, 196, 38, 10, 205, 230, 148, 186, 154, 20, 212, 109, 234, 47, 244, 42, 48, 8, 181, 197, 86, 21, 51, 105, 77, 211, 247, 190, 67, 168, 92, 253, 77, 161, 189]),
		timestamp: 200000,
	},
	{
		data: new Uint8Array([0, 0, 0, 11, 1, 158, 165, 68, 39, 255, 0, 0, 3, 1, 73]),
		timestamp: 166666,
	},
	{
		data: new Uint8Array([0, 0, 0, 19, 1, 158, 167, 68, 39, 255, 0, 15, 131, 178, 115, 244, 44, 8, 0, 112, 227, 114, 238]),
		timestamp: 233333,
	},
    ];

promise_test(async () => {
    const frameTimestamps = [];
    decoder =  new VideoDecoder({
        output(frame) {
            frameTimestamps.push(frame.timestamp);
            frame.close();
        },
        error(e) {
            console.log(e);
        }
    });

    decoder.configure({
        codec: 'avc1.640015',
        codedWidth: 480,
        codedHeight: 270,
        description: new Uint8Array([1, 100, 0, 21, 255, 225, 0, 26, 103, 100, 0, 21, 172, 217, 65, 224, 143, 235, 1, 16, 0, 0, 3, 0, 16, 0, 0, 3, 3, 192, 241, 98, 217, 96, 1, 0, 4, 104, 239, 139, 203]),
    });

    const chunks = frames.map((frame, i) => new EncodedVideoChunk({type: i == 0 ? 'key' : 'delta', timestamp: frame.timestamp, duration: 33333, data: frame.data}));

    for (let i = 0; i < chunks.length; ++i)
        decoder.decode(chunks[i]);

    await decoder.flush();

    assert_array_equals(frameTimestamps, frames.map(frame => frame.timestamp).sort((a,b) => a - b), "timestamps are ordered");
}, "H264 reordering with SPS containing emulation prevention byte");

promise_test(async () => {
    let frameTimestamps = [];
    decoder =  new VideoDecoder({
        output(frame) {
            frameTimestamps.push(frame.timestamp);
            frame.close();
        },
        error(e) {
            console.log(e);
        }
    });

    decoder.configure({
        codec: 'avc1.640015',
        codedWidth: 480,
        codedHeight: 270,
        description: new Uint8Array([1, 100, 0, 21, 255, 225, 0, 26, 103, 100, 0, 21, 172, 217, 65, 224, 143, 235, 1, 16, 0, 0, 3, 0, 16, 0, 0, 3, 3, 192, 241, 98, 217, 96, 1, 0, 4, 104, 239, 139, 203]),
    });

    // We change timetamps 33333 and 66666 to 333333 and 666666
    const computeTimestamp = (timestamp, i) => {
        if (i === 3)
            return 333333;
        if (i === 2)
            return 666666;
        return timestamp;
    };
    const chunks = frames.map((frame, i) => new EncodedVideoChunk({type: i == 0 ? 'key' : 'delta', timestamp: computeTimestamp(frame.timestamp, i), duration: 33333, data: frame.data}));

    // Expected frames ordering is governed by pic_order_count and not frame timestamp.
    let expectedTimestamps = frames.map(frame => frame.timestamp).sort((a,b) => a - b);
    const computeExpectedTimestamp = (timestamp, i) => {
        if (i === 1)
            return 333333;
        if (i === 2)
            return 666666;
        return timestamp;
    };
    expectedTimestamps = expectedTimestamps.map((timestamp, i) => computeExpectedTimestamp(timestamp, i));

    for (let i = 0; i < chunks.length; ++i)
        decoder.decode(chunks[i]);

    await decoder.flush();

    assert_array_equals(frameTimestamps, expectedTimestamps, "timestamps are ordered");
}, "H264 reordering should use pic order count");
</script>
    </script>
</body>
</html>
